pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tmux-cast"
version = "0.1.0"
description = "Stream tmux sessions to Chromecast devices"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["tmux", "chromecast", "streaming", "terminal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Terminals :: Terminal Emulators/X Terminals",
]
dependencies = [
    "pychromecast>=13.0.0",
    "pyte>=0.8.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
tmux-cast = "tmuxcast.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/tmux-cast"
Issues = "https://github.com/yourusername/tmux-cast/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
